modified: restapi project structure
